refactor(Templates): split article rules into more managable files
